+++ demo/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@iwsio/forms-demo",
-  "version": "4.2.1",
+  "version": "4.3.0-alpha.1",
   "private": true,
   "devDependencies": {
     "@babel/core": "^7.24.7",
@@ -8,40 +8,32 @@
     "@babel/preset-react": "^7.24.7",
     "@babel/preset-typescript": "^7.24.7",
     "@tailwindcss/typography": "^0.5.13",
-    "@tanstack/react-query": "^5.40.1",
-    "@testing-library/jest-dom": "^6.4.5",
+    "@tanstack/react-query": "^5.50.1",
+    "@testing-library/jest-dom": "^6.4.6",
     "@testing-library/react": "^16.0.0",
     "@testing-library/user-event": "^14.5.2",
-    "@types/node": "^20.14.2",
+    "@types/node": "^20.14.10",
     "@types/react": "^18.3.3",
     "@types/react-dom": "^18.3.0",
-    "@typescript-eslint/eslint-plugin": "^7.12.0",
-    "@typescript-eslint/parser": "^7.12.0",
     "autoprefixer": "^10.4.19",
     "babel-loader": "^9.1.3",
     "babel-plugin-prismjs": "^2.1.0",
     "css-loader": "^7.1.2",
-    "cssnano": "^7.0.2",
-    "daisyui": "^4.12.2",
-    "eslint": "^8.57.0",
-    "eslint-import-resolver-typescript": "^3.6.1",
-    "eslint-plugin-import": "^2.29.1",
-    "eslint-plugin-node": "^11.1.0",
-    "eslint-plugin-promise": "^6.2.0",
-    "eslint-plugin-react": "^7.34.2",
+    "cssnano": "^7.0.4",
+    "daisyui": "^4.12.10",
     "html-webpack-plugin": "^5.6.0",
     "jsdom": "^24.1.0",
     "markdown-it": "^14.1.0",
     "mini-css-extract-plugin": "^2.9.0",
     "npm-run-all": "^4.1.5",
-    "postcss": "^8.4.38",
+    "postcss": "^8.4.39",
     "postcss-cli": "^11.0.0",
     "postcss-js": "^4.0.1",
     "postcss-loader": "^8.1.1",
     "prismjs": "^1.29.0",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
-    "react-router-dom": "^6.23.1",
+    "react-router-dom": "^6.24.1",
     "rimraf": "^5.0.7",
     "source-map-loader": "^5.0.0",
     "style-loader": "^4.0.0",
@@ -49,9 +41,9 @@
     "terser-webpack-plugin": "^5.3.10",
     "tsconfig-paths-webpack-plugin": "^4.1.0",
     "tslib": "^2.6.3",
-    "typescript": "^5.4.5",
+    "typescript": "^5.5.3",
     "vitest": "^1.6.0",
-    "web-vitals": "^4.0.1",
+    "web-vitals": "^4.2.1",
     "webpack-bundle-analyzer": "^4.10.2",
     "webpack-cli": "^5.1.4",
     "webpack-dev-server": "^5.0.4"
