+++ demo/src/index.tsx
@@ -5,7 +5,7 @@ import reportWebVitals from './reportWebVitals'
 import './index.css'
 
 const root = ReactDOM.createRoot(
-  document.getElementById('root') as HTMLElement
+	document.getElementById('root') as HTMLElement
 )
 root.render(
 	<StrictMode>
