+++ demo/src/samples/RawSampleField.tsx
@@ -1,7 +1,7 @@
 import { FC, FormEventHandler, ReactNode, useState } from 'react'
 import { FieldManager } from '@iwsio/forms'
 
-export const RawSampleField: FC<{children?: ReactNode, title?: string, label?: string, help?: ReactNode}> = ({ children, title, label, help }) => {
+export const RawSampleField: FC<{ children?: ReactNode, title?: string, label?: string, help?: ReactNode }> = ({ children, title, label, help }) => {
 	const [success, setSuccess] = useState(false)
 
 	const handleSubmit = (_fields: Record<string, any>) => {
@@ -26,7 +26,10 @@ export const RawSampleField: FC<{children?: ReactNode, title?: string, label?: s
 
 				<div className="flex flex-row justify-end gap-3">
 					<button className="btn" type="reset">Reset</button>
-					<button className={`btn ${success ? 'btn-success' : ''}`} type="submit" onClick={() => setSuccess(false)}>{success && 'Success!!'}{!success && 'Submit'}</button>
+					<button className={`btn ${success ? 'btn-success' : ''}`} type="submit" onClick={() => setSuccess(false)}>
+						{success && 'Success!!'}
+						{!success && 'Submit'}
+					</button>
 				</div>
 			</fieldset>
 		</FieldManager>
