+++ demo/webpack.config.cjs
@@ -13,7 +13,7 @@ const OUTPUT_DIR = path.join(__dirname, './dist')
 
 function transformSharedToDefine(shared) {
 	const result = {}
-	Object.keys(shared).forEach(k => {
+	Object.keys(shared).forEach((k) => {
 		const v = shared[k]
 		result[`process.env.${k}`] = JSON.stringify(v)
 	})
@@ -39,7 +39,8 @@ function buildConfig(argv) {
 			enforce: 'pre',
 			use: ['source-map-loader']
 		}
-	} else {
+	}
+	else {
 		if (!CI) { // no bundle analyzer for a CI build
 			plugins = [new BundleAnalyzerPlugin({
 				openAnalyzer: false,
