--- eslint-node.json
@@ -1,18 +1,0 @@
-{
-	"extends": [
-		"./eslint-base.json",
-		"plugin:node/recommended"
-	],
-
-	"plugins": [
-		"node"
-	],
-
-	"rules": {
-		"node/no-missing-import": "off",
-		"node/no-unsupported-features/es-syntax": "off",
-		"node/no-extraneous-import": ["error", {
-			"allowModules": ["@iwsllc/ts-sync-shared"]
-		}]
-	}
-}
