--- forms/.eslintrc.json
@@ -1,3 +1,0 @@
-{
-	"extends": ["../eslint-react.json"]
-}
