+++ forms/src/Input.tsx
@@ -1,13 +1,10 @@
 import { ChangeEventHandler, InputHTMLAttributes, forwardRef, useEffect } from 'react'
-import { FieldChangeEventHandler, ValidationProps } from './types'
-import { useFieldManager } from './useFieldManager'
+import { ValidationProps } from './types'
 import { useForwardRef } from './useForwardRef'
 
 export type InputProps = ValidationProps & InputHTMLAttributes<HTMLInputElement>
 
-type Ref = HTMLInputElement;
-
-export type InputFieldProps = Omit<InputProps, 'DefaultValue' | 'onChange'> & { onChange?: FieldChangeEventHandler }
+export type Ref = HTMLInputElement
 
 export const Input = forwardRef<Ref, InputProps>(({ onFieldError, fieldError, name, type = 'text', onChange, value, checked, onInvalid, ...other }, ref) => {
 	const localRef = useForwardRef<Ref>(ref)
@@ -54,32 +51,3 @@ export const Input = forwardRef<Ref, InputProps>(({ onFieldError, fieldError, na
 	)
 })
 Input.displayName = 'Input'
-
-export const InputField = forwardRef<Ref, InputFieldProps>(({ type = 'text', name, onChange, value, ...other }, ref) => {
-	const localRef = useForwardRef<Ref>(ref)
-	const { handleChange: managerOnChange, fields, setFieldError, fieldErrors, mapError } = useFieldManager()
-
-	const handleOnChange: ChangeEventHandler<Ref> = (e) => {
-		const result = managerOnChange<HTMLInputElement>(e)
-		if (onChange != null) onChange(result)
-	}
-
-	const handleFieldError = (key, validity, message) => {
-		const mappedError = mapError(validity, message)
-		setFieldError(key, mappedError, validity)
-	}
-
-	return (
-		<Input
-			ref={localRef}
-			onChange={handleOnChange}
-			{...(/checkbox|radio/i.test(type) ? { value, checked: fields[name] === value } : { value: fields[name] })}
-			type={type}
-			name={name}
-			fieldError={fieldErrors[name]}
-			onFieldError={handleFieldError}
-			{...other}
-		/>
-	)
-})
-InputField.displayName = 'InputField'
