+++ forms/src/Select.tsx
@@ -1,13 +1,10 @@
 import { ChangeEventHandler, PropsWithChildren, SelectHTMLAttributes, forwardRef, useEffect } from 'react'
-import { FieldChangeEventHandler, ValidationProps } from './types'
-import { useFieldManager } from './useFieldManager'
+import { ValidationProps } from './types'
 import { useForwardRef } from './useForwardRef'
 
-export type SelectProps = PropsWithChildren & ValidationProps & SelectHTMLAttributes<HTMLSelectElement>
+export type SelectProps = PropsWithChildren<SelectHTMLAttributes<HTMLSelectElement>> & ValidationProps
 
-export type SelectFieldProps = Omit<SelectProps, 'DefaultValue' | 'onChange'> & { onChange?: FieldChangeEventHandler }
-
-type Ref = HTMLSelectElement;
+export type Ref = HTMLSelectElement
 
 export const Select = forwardRef<Ref, SelectProps>(({ onFieldError, onInvalid, fieldError, name, onChange, value, children, ...other }, ref) => {
 	const localRef = useForwardRef<Ref>(ref)
@@ -53,31 +50,3 @@ export const Select = forwardRef<Ref, SelectProps>(({ onFieldError, onInvalid, f
 	)
 })
 Select.displayName = 'Select'
-
-export const SelectField = forwardRef<Ref, Omit<SelectFieldProps, 'DefaultValue'>>(({ name, onChange, ...other }, ref) => {
-	const localRef = useForwardRef<Ref>(ref)
-	const { handleChange: managerOnChange, fields, setFieldError, fieldErrors, mapError } = useFieldManager()
-
-	const handleOnChange: ChangeEventHandler<Ref> = (e) => {
-		const result = managerOnChange(e)
-		if (onChange != null) onChange(result)
-	}
-
-	const handleFieldError = (key, validity, message) => {
-		const mappedError = mapError(validity, message)
-		setFieldError(key, mappedError, validity)
-	}
-
-	return (
-		<Select
-			ref={localRef}
-			onChange={handleOnChange}
-			onFieldError={handleFieldError}
-			name={name}
-			fieldError={fieldErrors[name]}
-			value={fields[name]}
-			{...other}
-		/>
-	)
-})
-SelectField.displayName = 'SelectField'
