+++ forms/src/TextArea.test.tsx
@@ -1,9 +1,7 @@
-import { useState } from 'react'
 import { act, render, screen } from '@testing-library/react'
 import userEvent from '@testing-library/user-event'
-import { TextArea, TextAreaField } from './TextArea'
-import { useFieldManager } from './useFieldManager'
-import { FieldManagerWrapper } from './__tests__/FieldManagerWrapper'
+import { useState } from 'react'
+import { TextArea } from './TextArea'
 import { FieldError } from './types'
 
 export const ControlledTextArea = () => {
@@ -19,7 +17,7 @@ export const ControlledTextAreaWithErrors = () => {
 	const [error, setError] = useState<FieldError | undefined>()
 	const handleChange = (e) => {
 		setValue(e.target.value)
-		if (e.target.value === 'abc') setError({ message: "Cannot enter 'abc'.", validity: { valid: false, customError: true } as any })
+		if (e.target.value === 'abc') setError({ message: 'Cannot enter \'abc\'.', validity: { valid: false, customError: true } as any })
 	}
 	const handleFieldError = (_key, validity, message) => {
 		setError({ message, validity })
@@ -33,7 +31,7 @@ export const ControlledTextAreaWithErrors = () => {
 	)
 }
 
-describe('TextArea', function() {
+describe('TextArea', function () {
 	it('should work as an uncontrolled textarea', async () => {
 		render(<TextArea name="field" required data-testid="field" />)
 
@@ -97,7 +95,7 @@ describe('TextArea', function() {
 
 		expect(textarea.validity.valid).to.be.false
 		expect(textarea.validity.customError).to.be.true
-		expect(textarea.validationMessage).to.eq("Cannot enter 'abc'.")
+		expect(textarea.validationMessage).to.eq('Cannot enter \'abc\'.')
 	})
 	it('should clear custom validity when upstream fieldError changes to nothing', async () => {
 		render(<ControlledTextAreaWithErrors />)
@@ -110,7 +108,7 @@ describe('TextArea', function() {
 
 		expect(textarea.validity.valid).to.be.false
 		expect(textarea.validity.customError).to.be.true
-		expect(textarea.validationMessage).to.eq("Cannot enter 'abc'.")
+		expect(textarea.validationMessage).to.eq('Cannot enter \'abc\'.')
 
 		await userEvent.click(screen.getByTestId('clear'))
 
@@ -119,94 +117,5 @@ describe('TextArea', function() {
 
 		expect(textarea.validity.valid).to.be.true
 		expect(textarea.validity.customError).to.be.false
-	})
-})
-
-describe('TextAreaField', () => {
-	it('should work as an controlled input and handle custom errors', async () => {
-		const CustomErrorField = () => {
-			const { setFieldError } = useFieldManager()
-			const handleChange = (e) => {
-				if (e.target.value === 'abc') setFieldError(e.target.name, "Cannot enter 'abc'.")
-			}
-			return <TextAreaField name="field" onChange={handleChange} required data-testid="field" />
-		}
-
-		render(<CustomErrorField />, { wrapper: FieldManagerWrapper })
-
-		expect(screen.getByTestId('field')).to.be.ok
-		const textarea = screen.getByTestId('field') as HTMLInputElement
-
-		// basic validation fail
-		act(() => { textarea.checkValidity() })
-
-		expect(textarea.validity.valid).to.be.false
-		expect(textarea.validity.valueMissing).to.be.true
-
-		await userEvent.clear(textarea)
-		await userEvent.type(textarea, 'ab')
-
-		// basic validation pass
-		expect(textarea.value).to.eq('ab')
-		expect(textarea.checkValidity()).to.be.true
-
-		await userEvent.type(textarea, 'c')
-
-		// validation fail (from controlled state error)
-		expect(textarea.value).to.eq('abc')
-
-		act(() => { textarea.checkValidity() })
-
-		expect(textarea.validity.valid).to.be.false
-		expect(textarea.validity.customError).to.be.true
-		expect(textarea.validationMessage).to.eq("Cannot enter 'abc'.")
-
-		await userEvent.type(textarea, 'c')
-
-		expect(textarea.value).to.eq('abcc')
-		expect(textarea.checkValidity()).to.be.true
-	})
-	it('should work as an controlled input and handle custom errors via field results', async () => {
-		const CustomErrorField = () => {
-			const { setFieldError } = useFieldManager()
-			const handleChange = (e) => {
-				if (e.fields.field === 'abc') setFieldError('field', "Cannot enter 'abc'.")
-			}
-			return <TextAreaField name="field" onChange={handleChange} required data-testid="field" />
-		}
-
-		render(<CustomErrorField />, { wrapper: FieldManagerWrapper })
-
-		expect(screen.getByTestId('field')).to.be.ok
-		const textarea = screen.getByTestId('field') as HTMLInputElement
-
-		// basic validation fail
-		act(() => { textarea.checkValidity() })
-
-		expect(textarea.validity.valid).to.be.false
-		expect(textarea.validity.valueMissing).to.be.true
-
-		await userEvent.clear(textarea)
-		await userEvent.type(textarea, 'ab')
-
-		// basic validation pass
-		expect(textarea.value).to.eq('ab')
-		expect(textarea.checkValidity()).to.be.true
-
-		await userEvent.type(textarea, 'c')
-
-		// validation fail (from controlled state error)
-		expect(textarea.value).to.eq('abc')
-
-		act(() => { textarea.checkValidity() })
-
-		expect(textarea.validity.valid).to.be.false
-		expect(textarea.validity.customError).to.be.true
-		expect(textarea.validationMessage).to.eq("Cannot enter 'abc'.")
-
-		await userEvent.type(textarea, 'c')
-
-		expect(textarea.value).to.eq('abcc')
-		expect(textarea.checkValidity()).to.be.true
 	})
 })
