+++ forms/src/TextArea.tsx
@@ -1,13 +1,10 @@
-import { forwardRef, useEffect, ChangeEventHandler, TextareaHTMLAttributes } from 'react'
+import { ChangeEventHandler, forwardRef, TextareaHTMLAttributes, useEffect } from 'react'
+import { ValidationProps } from './types'
 import { useForwardRef } from './useForwardRef'
-import { FieldChangeEventHandler, ValidationProps } from './types'
-import { useFieldManager } from './useFieldManager'
 
 export type TextAreaProps = ValidationProps & TextareaHTMLAttributes<HTMLTextAreaElement>
 
-export type TextAreaFieldProps = Omit<TextAreaProps, 'DefaultValue' | 'onChange'> & { onChange?: FieldChangeEventHandler }
-
-type Ref = HTMLTextAreaElement;
+export type Ref = HTMLTextAreaElement
 
 export const TextArea = forwardRef<Ref, TextAreaProps>(({ onFieldError, onInvalid, fieldError, name, onChange, value, ...other }, ref) => {
 	const localRef = useForwardRef<Ref>(ref)
@@ -50,32 +47,5 @@ export const TextArea = forwardRef<Ref, TextAreaProps>(({ onFieldError, onInvali
 		/>
 	)
 })
+
 TextArea.displayName = 'TextArea'
-
-export const TextAreaField = forwardRef<Ref, Omit<TextAreaFieldProps, 'DefaultValue'>>(({ name, onChange, ...other }, ref) => {
-	const localRef = useForwardRef<Ref>(ref)
-
-	const { handleChange: managerOnChange, fields, setFieldError, fieldErrors, mapError } = useFieldManager()
-
-	const handleOnChange: ChangeEventHandler<Ref> = (e) => {
-		const result = managerOnChange(e)
-		if (onChange != null) onChange(result)
-	}
-	const handleFieldError = (key, validity, message) => {
-		const mappedError = mapError(validity, message)
-		setFieldError(key, mappedError, validity)
-	}
-
-	return (
-		<TextArea
-			ref={localRef}
-			onChange={handleOnChange}
-			onFieldError={handleFieldError}
-			name={name}
-			fieldError={fieldErrors[name]}
-			value={fields[name]}
-			{...other}
-		/>
-	)
-})
-TextAreaField.displayName = 'TextAreaField'
