+++ forms/src/index.ts
@@ -1,9 +1,13 @@
-export * from './Input'
-export * from './Select'
-export * from './TextArea'
+export { Input, InputProps } from './Input'
+export * from './InputField'
+export { Select, SelectProps } from './Select'
+export * from './SelectField'
+export { TextArea, TextAreaProps } from './TextArea'
+export * from './TextAreaField'
 export * from './InvalidFeedbackForField'
 export * from './types'
 export * from './FieldManager'
+export * from './ControlledFieldManager'
 export * from './useFieldState'
 export * from './useFieldManager'
 export * from './useForwardRef'
