+++ forms/src/useForwardRef.ts
@@ -2,7 +2,7 @@
 // Thank you https://github.com/lejahmie
 import { ForwardedRef, useEffect, useRef } from 'react'
 
-export const useForwardRef = <T, >(
+export const useForwardRef = <T >(
 	ref: ForwardedRef<T>,
 	initialValue: any = null
 ) => {
@@ -13,7 +13,8 @@ export const useForwardRef = <T, >(
 
 		if (typeof ref === 'function') {
 			ref(targetRef.current)
-		} else {
+		}
+		else {
 			ref.current = targetRef.current
 		}
 	}, [ref])
