+++ forms/tsconfig.build.json
@@ -1,8 +1,9 @@
 {
 	"extends": "./tsconfig",
 	"compilerOptions": {
-		"sourceMap": false,
-		"declarationMap": false
+		"sourceMap": true,
+		"declarationMap": false,
+		"inlineSources": true
 	},
 	"exclude": ["node_modules", "**/*.test.*", "**/__tests__/*", "**/__mocks__/*"]
 }
