+++ package.json
@@ -1,10 +1,10 @@
 {
   "name": "@iwsio/forms-root",
-  "version": "4.2.1",
+  "version": "4.3.0-alpha.1",
   "description": "Simple library with useful React forms components and browser validation.",
   "files": [],
   "engines": {
-    "node": ">=20"
+    "node": "22"
   },
   "repository": {
     "type": "git",
@@ -40,6 +40,12 @@
     "demo"
   ],
   "devDependencies": {
-    "demo": "file:demo"
+    "@eslint/js": "^9.6.0",
+    "@stylistic/eslint-plugin": "^2.3.0",
+    "eslint": "^8.57.0",
+    "eslint-plugin-promise": "^6.4.0",
+    "eslint-plugin-react": "^7.34.3",
+    "globals": "^15.8.0",
+    "typescript-eslint": "^7.15.0"
   }
 }
