+++ tsconfig-base.json
@@ -17,7 +17,7 @@
 		"skipDefaultLibCheck": true,
 		"skipLibCheck": true,
 		"sourceMap": true,
-		"target": "ES2015",
+		"target": "es2022",
 		"types": ["vitest/globals"]
 	}
 }
